# api/openapi.yaml
openapi: 3.1.0
info:
  title: Zoomies API
  version: "1.0.0"
  summary: Self-hosted GitHub Actions runner fleet controller
  description: |
    Zoomies manages ephemeral self-hosted GitHub Actions runners.

    This document is the whole product surface. The web UI and the `zoomies`
    CLI are both clients of it, and nothing is reachable from either that is
    not described here.

    ## Authentication

    Two credentials are accepted on the same routes:

    * a **session cookie** (`zoomies_session`), which the browser gets from
      `POST /auth/login`;
    * a **bearer token** (`Authorization: Bearer zoo_...`), for automation.

    Cookie-authenticated mutating requests are subject to a same-origin check.
    Bearer-authenticated requests are not, because they are not vulnerable to
    CSRF.

    Agents use a separate credential class on `/agent/*` and can reach nothing
    else.

    ## Roles

    | Role | May |
    |---|---|
    | `viewer` | read everything except secrets |
    | `operator` | act on the fleet and manage pools |
    | `admin` | manage users, tokens, installations and settings |

    Each operation below records its minimum role in `x-zoomies-role`.

    ## Conventions

    * List endpoints take `limit`, `offset`, `sort` and `order`, and return
      `{ items, total, limit, offset }`.
    * Timestamps are RFC 3339 with a `Z` offset.
    * Durations are Go duration strings: `"5m"`, `"1h30s"`.
...
      message written for a human to read.
  license:
...
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /auth/bootstrap:
    post:
      operationId: bootstrap
      tags: [auth]
      summary: Create the first administrator
      description: |
        Available only while no user exists. Once any account has been created
        this returns 409 forever. That check is the entire security of this
        route, which is why it is unauthenticated.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, minLength: 1, maxLength: 64, examples: [admin] }
                password: { type: string, minLength: 12, format: password }
                email: { type: string, format: email }
      responses:
        "201":
          description: Created, and the session cookie is set
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Identity" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with a username and password
      description: Rate limited per source address. A wrong password and an unknown user are indistinguishable, deliberately.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Identity" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: The current identity
      x-zoomies-role: viewer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Identity" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/password:
    post:
      operationId: changeOwnPassword
      tags: [auth]
      summary: Change your own password
      description: Invalidates every other session belonging to this account.
      x-zoomies-role: viewer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_password]
              properties:
                old_password: { type: string, format: password }
                new_password: { type: string, minLength: 12, format: password }
      responses:
        "204": { description: Changed }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/Unprocessable" }

  # ==========================================================================
  # Overview
  # ==========================================================================
  /stats:
    get:
      operationId: getStats
      tags: [overview]
      summary: Fleet statistics
      x-zoomies-role: viewer
      parameters:
        - name: window
          in: query
          description: Rolling window for the completed/failed counts and queue-wait percentiles.
          schema: { type: string, default: "1h", examples: ["1h", "24h"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Stats" }
  /samples:
    get:
      operationId: listSamples
      tags: [overview]
      summary: Fleet samples for the sparklines
      x-zoomies-role: viewer
      parameters:
        - name: window
          in: query
          schema: { type: string, default: "1h" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/FleetSample" }
...
  /scaling-events:
    get:
      operationId: listScalingEvents
      tags: [overview]
      summary: Recent scaling decisions, with the reason in words
      x-zoomies-role: viewer
      parameters:
        - name: pool_id
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ScalingEvent" }
  /events:
    get:
      operationId: streamEvents
      tags: [overview]
      summary: Live event stream (Server-Sent Events)
      description: |
        `text/event-stream`. Every state change the operator would want to see.
        Honours `Last-Event-ID` on reconnect and replays what the server still
        holds. A `heartbeat` event every 20 seconds keeps proxies from closing
        an idle connection.
      x-zoomies-role: viewer
      parameters:
        - name: kinds
          in: query
          description: Comma-separated event kinds to receive. Omit for all.
          schema: { type: string, examples: ["runner.updated,scaling"] }
        - name: topic
          in: query
          description: Topic prefix filter, e.g. `runner:run_k3f9qz2m`.
          schema: { type: string }
      responses:
        "200":
          description: The stream
          content:
            text/event-stream:
              schema: { type: string }

  # ==========================================================================
  # Installations
  # ==========================================================================
  /installations:
    get:
      operationId: listInstallations
      tags: [installations]
      summary: List GitHub App installations
      x-zoomies-role: viewer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Installation" }
    post:
      operationId: createInstallation
      tags: [installations]
      summary: Connect a GitHub App installation
      description: The private key is sealed with the instance encryption key before it reaches the database, and is never returned.
      x-zoomies-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InstallationCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Installation" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /installations/{id}:
    parameters: [{ $ref: "#/components/parameters/PathID" }]
    get:
      operationId: getInstallation
      tags: [installations]
      summary: Get an installation
      x-zoomies-role: viewer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Installation" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateInstallation
      tags: [installations]
      summary: Update an installation
      x-zoomies-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InstallationUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Installation" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteInstallation
      tags: [installations]
      summary: Remove an installation
      description: Cascades to its pools. The response says how many pools and runners went with it.
      x-zoomies-role: admin
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  pools_deleted: { type: integer }
                  runners_affected: { type: integer }
        "404": { $ref: "#/components/responses/NotFound" }
  /installations/{id}/verify:
    parameters: [{ $ref: "#/components/parameters/PathID" }]
    post:
      operationId: verifyInstallation
      tags: [installations]
      summary: Check the credentials and permissions
      description: On a 403 from GitHub, the message names the permission that is probably missing rather than echoing the status code.
      x-zoomies-role: operator
      responses:
        "200":
          description: The result of the probe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstallationHealth" }
        "404": { $ref: "#/components/responses/NotFound" }
  /installations/{id}/runner-groups:
    parameters: [{ $ref: "#/components/parameters/PathID" }]
    get:
      operationId: listRunnerGroups
      tags: [installations]
      summary: The target's runner groups
      x-zoomies-role: viewer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer, format: int64 }
                        name: { type: string }
  /installations/{id}/rate-limit:
    parameters: [{ $ref: "#/components/parameters/PathID" }]
    get:
      operationId: getRateLimit
      tags: [installations]
      summary: Remaining GitHub API quota
      x-zoomies-role: viewer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  limit: { type: integer }
                  remaining: { type: integer }
                  reset_at: { type: string, format: date-time }
  /installations/manifest:
    post:
      operationId: createAppManifest
      tags: [installations]
      summary: Build a GitHub App manifest
      description: |
        Returns the manifest JSON and the URL to POST it to. The browser submits
        it, the operator confirms in GitHub, and GitHub redirects back with a
        code that `manifest/exchange` turns into App credentials.

        The manifest asks for exactly the permissions Zoomies needs and no more.
      x-zoomies-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target, target_type]
              properties:
                name: { type: string, examples: ["Zoomies (acme)"] }
                target: { type: string, examples: [acme] }
                target_type: { $ref: "#/components/schemas/TargetType" }
                api_base_url: { type: string, examples: ["https://api.github.com"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  post_url: { type: string, description: Where the browser POSTs the manifest. }
                  manifest: { type: string, description: The manifest JSON, to be sent as a form field named `manifest`. }
                  state: { type: string, description: Single-use value echoed back on the callback. }
  /installations/manifest/exchange:
    post:
      operationId: exchangeAppManifest
      tags: [installations]
      summary: Exchange a manifest code for App credentials
      x-zoomies-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
                state: { type: string }
                api_base_url: { type: string }
      responses:
        "200":
          description: The App was created; finish by installing it.
          content:
            application/json:
              schema:
                type: object
                properties:
                  app_id: { type: integer, format: int64 }
                  slug: { type: string }
                  name: { type: string }
                  html_url: { type: string }
                  install_url: { type: string, description: Send the operator here to install the App on their organisation. }
        "422": { $ref: "#/components/responses/Unprocessable" }
...
  /pools:
    get:
      operationId: listPools
      tags: [pools]
      summary: List pools
      description: Each pool carries live per-state runner counts and its current utilisation.
      x-zoomies-role: viewer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Pool" }
    post:
      operationId: createPool
      tags: [pools]
      summary: Create a pool
      x-zoomies-role: operator
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PoolCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pool" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /pools/validate:
    post:
      operationId: validatePool
      tags: [pools]
      summary: Dry-run a pool definition
      description: |
        Returns the field errors and the dangerous-setting warnings the pool
        would produce, without creating anything. The creation wizard's review
        step calls this, so the warnings an operator sees before confirming are
        the same ones the server would raise afterwards.
      x-zoomies-role: operator
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PoolCreate" }
      responses:
        "200":
          description: The verdict
          content:
            application/json:
              schema:
                type: object
                required: [valid]
                properties:
                  valid: { type: boolean }
                  errors:
                    type: array
                    items: { $ref: "#/components/schemas/FieldError" }
                  warnings:
                    type: array
                    items: { $ref: "#/components/schemas/Problem" }
                  matching_hosts:
                    type: integer
                    description: How many hosts could actually run this pool. Zero is worth saying out loud before the pool is created.
...
  /pools/{id}/enable:
    parameters: [{ $ref: "#/components/parameters/PathID" }]
    post:
      operationId: enablePool
      tags: [pools]
      summary: Enable a pool
      x-zoomies-role: operator
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pool" }
  /pools/{id}/disable:
    parameters: [{ $ref: "#/components/parameters/PathID" }]
    post:
      operationId: disablePool
      tags: [pools]
      summary: Disable a pool
      description: Existing runners drain as they become idle; no new ones are created. Running jobs are not interrupted.
      x-zoomies-role: operator
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pool" }

  # ==========================================================================
  # Runners
  # ==========================================================================
...
  /tokens/{id}:
    parameters: [{ $ref: "#/components/parameters/PathID" }]
    delete:
      operationId: revokeToken
      tags: [admin]
      summary: Revoke an API token
      x-zoomies-role: admin
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: zoomies_session
      description: Set by `POST /auth/login`. HttpOnly, SameSite=Lax, and Secure when the external URL is https.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: zoo_<prefix>_<secret>
      description: An API token. Not subject to the same-origin check, because it is not vulnerable to CSRF.
...